+++ src/components/content/blockquote/blockquote.styles.ts
@@ -1,11 +1,13 @@
 import { css, unsafeCSS } from 'lit';
 import { breakpoints } from '../../../assets/styles/breakpoints.js';
-import { boxSizingReset, slottedReset, inheritedTextReset } from '../../../assets/styles/style-resets.js';
+import { slottedReset, inheritedTextReset } from '../../../assets/styles/style-resets.js';
 
 const smMax = unsafeCSS(breakpoints.smMax);
 
 export const blockquoteStyles = css`
-	${boxSizingReset}
+	:host {
+		box-sizing: border-box;
+	}
 
 
 	/* # Host */