+++ src/components/forms/form-field/form-field.styles.ts
@@ -1,11 +1,13 @@
 import { css, unsafeCSS } from 'lit';
 import { breakpoints } from '../../../assets/styles/breakpoints.js';
-import { boxSizingReset, slottedReset, inheritedTextReset } from '../../../assets/styles/style-resets.js';
+import { slottedReset, inheritedTextReset } from '../../../assets/styles/style-resets.js';
 
 const mdMin = unsafeCSS(breakpoints.mdMin);
 
 export const formFieldStyles = css`
-	${boxSizingReset}
+	:host {
+		box-sizing: border-box;
+	}
 
 
 	/* # Host */