+++ src/components/inputs/checkbox/checkbox.styles.ts
@@ -1,8 +1,9 @@
 import { css } from 'lit';
-import { boxSizingReset } from '../../../assets/styles/style-resets.js';
 
 export const checkboxStyles = css`
-	${boxSizingReset}
+	:host {
+		box-sizing: border-box;
+	}
 
 
 	/* # Host */