+++ src/components/layout/page-sections/hero/hero.styles.ts
@@ -1,6 +1,6 @@
 import { css, unsafeCSS } from 'lit';
 import { breakpoints } from '../../../../assets/styles/breakpoints.js';
-import { boxSizingReset, inheritedTextReset, slottedReset } from '../../../../assets/styles/style-resets.js';
+import { inheritedTextReset, slottedReset } from '../../../../assets/styles/style-resets.js';
 
 const smMax = unsafeCSS(breakpoints.smMax);
 const mdMin = unsafeCSS(breakpoints.mdMin);
@@ -8,7 +8,9 @@ const mdMax = unsafeCSS(breakpoints.mdMax);
 const lgMin = unsafeCSS(breakpoints.lgMin);
 
 export const heroStyles = css`
-	${boxSizingReset}
+	:host {
+		box-sizing: border-box;
+	}
 
 
 	/* # Host */