+++ src/components/layout/split-views/side-by-side-split-view/side-by-side-split-view.styles.ts
@@ -1,8 +1,9 @@
 import { css } from 'lit';
-import { boxSizingReset } from '../../../../assets/styles/style-resets.js';
 
 export const sideBySideSplitViewStyles = css`
-	${boxSizingReset}
+	:host {
+		box-sizing: border-box;
+	}
 
 
 	/* # Host */