+++ src/components/lists-and-tables/table/table.styles.ts
@@ -1,8 +1,10 @@
 import { css } from 'lit';
-import { boxSizingReset, inheritedTextReset } from '../../../assets/styles/style-resets.js';
+import { inheritedTextReset } from '../../../assets/styles/style-resets.js';
 
 export const tableStyles = css`
-	${boxSizingReset}
+	:host {
+		box-sizing: border-box;
+	}
 	:host {
 		--_columns: none;
 		--_corner-radius: var(--semantics-tables-corner-radius);