+++ src/components/navigation/pagination/pagination.styles.ts
@@ -1,8 +1,10 @@
 import { css } from 'lit';
-import { boxSizingReset, inheritedTextReset } from '../../../assets/styles/style-resets.js';
+import { inheritedTextReset } from '../../../assets/styles/style-resets.js';
 
 export const paginationStyles = css`
-	${boxSizingReset}
+	:host {
+		box-sizing: border-box;
+	}
 
 
 	/* # Host */